refactor(docs): image dx
