+++ apps/docs/content/components/image/blurred.raw.jsx
@@ -1,0 +1,13 @@
+import {Image} from "@nextui-org/react";
+
+export default function App() {
+  return (
+    <Image
+      isBlurred
+      alt="NextUI Album Cover"
+      className="m-5"
+      src="https://nextui.org/images/album-cover.png"
+      width={240}
+    />
+  );
+}
