+++ apps/docs/content/components/image/blurred.ts
@@ -1,16 +1,4 @@
-const App = `import {Image} from "@nextui-org/react";
-
-export default function App() {
-  return (
-    <Image
-      isBlurred
-      width={240}
-      src="https://nextui.org/images/album-cover.png"
-      alt="NextUI Album Cover"
-      className="m-5"
-    />
-  );
-}`;
+import App from "./blurred.raw.jsx?raw";
 
 const react = {
   "/App.jsx": App,
