+++ apps/docs/content/components/image/loading.raw.jsx
@@ -1,0 +1,12 @@
+import {Image} from "@nextui-org/react";
+
+export default function App() {
+  return (
+    <Image
+      alt="NextUI hero Image with delay"
+      height={200}
+      src="https://app.requestly.io/delay/5000/https://nextui.org/images/hero-card-complete.jpeg"
+      width={300}
+    />
+  );
+}
