+++ apps/docs/content/components/image/usage.raw.jsx
@@ -1,0 +1,11 @@
+import {Image} from "@nextui-org/react";
+
+export default function App() {
+  return (
+    <Image
+      alt="NextUI hero Image"
+      src="https://nextui.org/images/hero-card-complete.jpeg"
+      width={300}
+    />
+  );
+}
