+++ apps/docs/content/components/image/zoomed.raw.jsx
@@ -1,0 +1,12 @@
+import {Image} from "@nextui-org/react";
+
+export default function App() {
+  return (
+    <Image
+      isZoomed
+      alt="NextUI Fruit Image with Zoom"
+      src="https://nextui.org/images/fruit-1.jpeg"
+      width={240}
+    />
+  );
+}
