+++ apps/docs/content/components/image/zoomed.ts
@@ -1,15 +1,4 @@
-const App = `import {Image} from "@nextui-org/react";
-
-export default function App() {
-  return (
-    <Image
-      isZoomed
-      width={240}
-      alt="NextUI Fruit Image with Zoom"
-      src="https://nextui.org/images/fruit-1.jpeg"
-    />
-  );
-}`;
+import App from "./zoomed.raw.jsx?raw";
 
 const react = {
   "/App.jsx": App,
